pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gh-emu-migration"
version = "0.1.0"
description = "POC: GitHub Enterprise ADFS→Entra ID SSO + EMU Migration Tool"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31,<3",
    "pyyaml>=6.0,<7",
    "rich>=13.0,<14",
    "click>=8.1,<9",
    "pydantic>=2.0,<3",
    "pywebview>=5.0,<6",
]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-cov", "ruff"]
packaging = ["pyinstaller>=6.0"]

[project.scripts]
emu-migrate = "emu_migration.cli:main"
emu-migrate-desktop = "emu_migration.desktop:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
emu_migration = ["ui/*.html", "ui/*.css", "ui/*.js"]

[tool.ruff]
target-version = "py310"
line-length = 120
